Convert market ID type to i32
Market IDs (markets.id, markets.owner_id) are INTEGER (int4) in Postgres.
Switch Market.id/owner_id and the MarketStore id parameters to i32 and drop
the compensating CAST(... AS BIGINT) casts. Cross-store calls into the
still-i64 instrument and position stores get a temporary .into().

+++ bot/src/ui.rs
@@ -68,7 +68,7 @@ pub fn get_command_option_user<'i>(command: &'i CommandInteraction, name: &str) 
     get_command_option_value(command, name).and_then(|o| o.as_user_id())
 }
 
-pub fn format_market_id(id: i64) -> String {
+pub fn format_market_id(id: i32) -> String {
     format!("#{id:04}")
 }
